Refactored to use promises, updated eslint & editorconfig to use Mozilla.org standard.
