+++ .editorconfig
@@ -1,3 +1,6 @@
+# EditorConfig is awesome: http://EditorConfig.org
+
+# top-most EditorConfig file
 root = true
 
 # Unix-style newlines with a newline ending every file
@@ -5,8 +8,32 @@ root = true
 end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
-
-# Indentiation
-[*.{css,html,js,json,py,styl,php}]
 indent_style = space
+
+[*.py]
 indent_size = 4
+
+[*.js]
+indent_size = 4
+
+[*.json]
+indent_size = 4
+
+[*.css]
+indent_size = 4
+
+[*.less]
+indent_size = 4
+
+[*.scss]
+indent_size = 4
+
+[*.html]
+indent_size = 2
+
+# Matches the exact files package.json and .travis.yml
+[{package.json, .travis.yml}]
+indent_size = 2
+
+[Makefile]
+indent_style = tab
