+++ .eslintrc.js
@@ -2,6 +2,7 @@ module.exports = {
     "env": {
         "node": true
     },
+    //"extends": "eslint:recommended",
     "parserOptions": {
         "ecmaVersion": 6,
         "sourceType": "module",
@@ -22,6 +23,25 @@ module.exports = {
         "semi": [
             2,
             "always"
-        ]
+        ],
+        "curly": [
+            2,
+            "all"
+        ],
+        "camelcase": [
+            2,
+            {
+                "properties": "always"
+            }
+        ],
+        "eqeqeq": [
+            2,
+            "smart"
+        ],
+        "one-var-declaration-per-line": [
+            2,
+            "always"
+        ],
+        "new-cap": 2
     }
 };
